+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Update openapparel.org references [#2281](https://github.com/open-apparel-registry/open-apparel-registry/pull/2281)
+
 ### Deprecated
 
 ### Removed